+++ doc/WhatsNew.rst
@@ -2,6 +2,15 @@
 What's New in Ginga?
 ++++++++++++++++++++
 
+Ver 2.0.20140114070809
+======================
+- The SAMP plugin has been updated to work with the new astropy.vo.samp
+module.
+- The Catalogs plugin has been updated to allow the user to define the
+radius of the conesearch or image search by drawing a circle (as well as
+the previous option--a rectangle).
+
+
 Ver 2.0.20131218034517
 ======================
 The user interface mapping just got a bit easier to use.  Ginga now